+++ typescript/infra/config/environments/testnet4/agent.ts
@@ -2,6 +2,7 @@ import {
   GasPaymentEnforcementPolicyType,
   RpcConsensusType,
   chainMetadata,
+  getDomainId,
 } from '@hyperlane-xyz/sdk';
 
 import {
@@ -49,7 +50,7 @@ const hyperlane: RootAgentConfig = {
     rpcConsensusType: RpcConsensusType.Fallback,
     docker: {
       repo,
-      tag: '5423057-20240103-221936',
+      tag: 'c9ef41c-20240126-172723',
     },
     blacklist: [
       ...releaseCandidateHelloworldMatchingList,
@@ -59,14 +60,40 @@ const hyperlane: RootAgentConfig = {
         // gas, we blacklist the old TestRecipient address.
         recipientAddress: '0xBC3cFeca7Df5A45d61BC60E7898E63670e1654aE',
       },
+      // OptimismGoerli griefers:
+      {
+        destinationDomain: getDomainId(chainMetadata.optimismgoerli),
+        recipientAddress: [
+          '0xed4de02c6f4cb1161bdfefdb2fcdeef4546fa36c',
+          '0x723192fc414fe536b414117a4b2c5a7b71f912e3',
+          '0x5a48723d80a7ee3be6855ca293059b5287ee6689',
+        ],
+      },
+      // Goerli griefers:
+      {
+        destinationDomain: getDomainId(chainMetadata.goerli),
+        recipientAddress: [
+          '0x0461c69ff7f29cfb5efd36b9d377fdfc95418c2b',
+          '0xe747c82ed8560ba137b24a3a97ff7504b50c3e91',
+          '0x6ad92511ee4a3835bde9b1bfd7063023b56a8c56',
+        ],
+      },
     ],
     gasPaymentEnforcement,
+    metricAppContexts: [
+      {
+        name: 'helloworld',
+        matchingList: routerMatchingList(
+          helloWorld[Contexts.Hyperlane].addresses,
+        ),
+      },
+    ],
   },
   validators: {
     rpcConsensusType: RpcConsensusType.Fallback,
     docker: {
       repo,
-      tag: '34611f0-20231218-204504',
+      tag: 'c9ef41c-20240126-172723',
     },
     chains: validatorChainConfig(Contexts.Hyperlane),
   },
@@ -74,7 +101,7 @@ const hyperlane: RootAgentConfig = {
     rpcConsensusType: RpcConsensusType.Fallback,
     docker: {
       repo,
-      tag: '5423057-20240103-221936',
+      tag: '8ccfdb7-20240103-084118',
     },
   },
 };
@@ -87,7 +114,7 @@ const releaseCandidate: RootAgentConfig = {
     rpcConsensusType: RpcConsensusType.Fallback,
     docker: {
       repo,
-      tag: '34611f0-20231218-204504',
+      tag: 'c9ef41c-20240126-172723',
     },
     whitelist: [...releaseCandidateHelloworldMatchingList],
     gasPaymentEnforcement,
@@ -100,7 +127,7 @@ const releaseCandidate: RootAgentConfig = {
     rpcConsensusType: RpcConsensusType.Fallback,
     docker: {
       repo,
-      tag: '34611f0-20231218-204504',
+      tag: 'c9ef41c-20240126-172723',
     },
     chains: validatorChainConfig(Contexts.ReleaseCandidate),
   },