+++ typescript/infra/config/environments/testnet4/validators.ts
@@ -1,4 +1,4 @@
-import { chainMetadata } from '@hyperlane-xyz/sdk';
+import { chainMetadata, getReorgPeriod } from '@hyperlane-xyz/sdk';
 
 import { ValidatorBaseChainConfigMap } from '../../../src/config/agent';
 import { Contexts } from '../../contexts';
@@ -13,7 +13,7 @@ export const validatorChainConfig = (
   return {
     alfajores: {
       interval: 5,
-      reorgPeriod: chainMetadata.alfajores.blocks!.reorgPeriod!,
+      reorgPeriod: getReorgPeriod(chainMetadata.alfajores),
       validators: validatorsConfig(
         {
           [Contexts.Hyperlane]: [
@@ -45,7 +45,7 @@ export const validatorChainConfig = (
     },
     basegoerli: {
       interval: 5,
-      reorgPeriod: chainMetadata.basegoerli.blocks!.reorgPeriod!,
+      reorgPeriod: getReorgPeriod(chainMetadata.alfajores),
       validators: validatorsConfig(
         {
           [Contexts.Hyperlane]: [
@@ -65,7 +65,7 @@ export const validatorChainConfig = (
     },
     fuji: {
       interval: 5,
-      reorgPeriod: chainMetadata.fuji.blocks!.reorgPeriod!,
+      reorgPeriod: getReorgPeriod(chainMetadata.alfajores),
       validators: validatorsConfig(
         {
           [Contexts.Hyperlane]: [
@@ -85,7 +85,7 @@ export const validatorChainConfig = (
     },
     // chiado: {
     //   interval: 5,
-    //   reorgPeriod: chainMetadata.chiado.blocks!.reorgPeriod!,
+    //   reorgPeriod: getReorgPeriod(chainMetadata.chiado),
     //   validators: validatorsConfig(
     //     {
     //       [Contexts.Hyperlane]: [
@@ -104,7 +104,7 @@ export const validatorChainConfig = (
     // },
     // lineagoerli: {
     //   interval: 5,
-    //   reorgPeriod: chainMetadata.lineagoerli.blocks!.reorgPeriod!,
+    //   reorgPeriod: getReorgPeriod(chainMetadata.lineagoerli),
     //   validators: validatorsConfig(
     //     {
     //       [Contexts.Hyperlane]: [
@@ -123,7 +123,7 @@ export const validatorChainConfig = (
     // },
     mumbai: {
       interval: 5,
-      reorgPeriod: chainMetadata.mumbai.blocks!.reorgPeriod!,
+      reorgPeriod: getReorgPeriod(chainMetadata.mumbai),
       validators: validatorsConfig(
         {
           [Contexts.Hyperlane]: [
@@ -143,7 +143,7 @@ export const validatorChainConfig = (
     },
     bsctestnet: {
       interval: 5,
-      reorgPeriod: chainMetadata.bsctestnet.blocks!.reorgPeriod!,
+      reorgPeriod: getReorgPeriod(chainMetadata.bsctestnet),
       validators: validatorsConfig(
         {
           [Contexts.Hyperlane]: [
@@ -163,7 +163,7 @@ export const validatorChainConfig = (
     },
     goerli: {
       interval: 5,
-      reorgPeriod: chainMetadata.goerli.blocks!.reorgPeriod!,
+      reorgPeriod: getReorgPeriod(chainMetadata.goerli),
       validators: validatorsConfig(
         {
           [Contexts.Hyperlane]: [
@@ -183,7 +183,7 @@ export const validatorChainConfig = (
     },
     scrollsepolia: {
       interval: 5,
-      reorgPeriod: chainMetadata.scrollsepolia.blocks!.reorgPeriod!,
+      reorgPeriod: getReorgPeriod(chainMetadata.scrollsepolia),
       validators: validatorsConfig(
         {
           [Contexts.Hyperlane]: [
@@ -203,7 +203,7 @@ export const validatorChainConfig = (
     },
     sepolia: {
       interval: 5,
-      reorgPeriod: chainMetadata.sepolia.blocks!.reorgPeriod!,
+      reorgPeriod: getReorgPeriod(chainMetadata.sepolia),
       validators: validatorsConfig(
         {
           [Contexts.Hyperlane]: [
@@ -223,7 +223,7 @@ export const validatorChainConfig = (
     },
     moonbasealpha: {
       interval: 5,
-      reorgPeriod: chainMetadata.moonbasealpha.blocks!.reorgPeriod!,
+      reorgPeriod: getReorgPeriod(chainMetadata.moonbasealpha),
       validators: validatorsConfig(
         {
           [Contexts.Hyperlane]: [
@@ -243,7 +243,7 @@ export const validatorChainConfig = (
     },
     optimismgoerli: {
       interval: 5,
-      reorgPeriod: chainMetadata.optimismgoerli.blocks!.reorgPeriod!,
+      reorgPeriod: getReorgPeriod(chainMetadata.optimismgoerli),
       validators: validatorsConfig(
         {
           [Contexts.Hyperlane]: [
@@ -263,7 +263,7 @@ export const validatorChainConfig = (
     },
     arbitrumgoerli: {
       interval: 5,
-      reorgPeriod: chainMetadata.arbitrumgoerli.blocks!.reorgPeriod!,
+      reorgPeriod: getReorgPeriod(chainMetadata.arbitrumgoerli),
       validators: validatorsConfig(
         {
           [Contexts.Hyperlane]: [
@@ -283,7 +283,7 @@ export const validatorChainConfig = (
     },
     polygonzkevmtestnet: {
       interval: 5,
-      reorgPeriod: chainMetadata.polygonzkevmtestnet.blocks!.reorgPeriod!,
+      reorgPeriod: getReorgPeriod(chainMetadata.polygonzkevmtestnet),
       validators: validatorsConfig(
         {
           [Contexts.Hyperlane]: [
@@ -303,7 +303,7 @@ export const validatorChainConfig = (
     },
     // proteustestnet: {
     //   interval: 5,
-    //   reorgPeriod: chainMetadata.proteustestnet.blocks!.reorgPeriod!,
+    //   reorgPeriod: getReorgPeriod(chainMetadata.proteustestnet),
     //   validators: validatorsConfig(
     //     {
     //       [Contexts.Hyperlane]: [
@@ -318,22 +318,29 @@ export const validatorChainConfig = (
     //     'proteustestnet',
     //   ),
     // },
-    // solanadevnet: {
-    //   interval: 10,
-    //   reorgPeriod: chainMetadata.solanadevnet.blocks!.reorgPeriod!,
-    //   validators: validatorsConfig(
-    //     {
-    //       [Contexts.Hyperlane]: [
-    //         '0xec0f73dbc5b1962a20f7dcbe07c98414025b0c43',
-    //         '0x9c20a149dfa09ea9f77f5a7ca09ed44f9c025133',
-    //         '0x967c5ecdf2625ae86580bd203b630abaaf85cd62',
-    //       ],
-    //       [Contexts.ReleaseCandidate]: [
-    //         '0x21b9eff4d1a6d3122596c7fb80315bf094b6e5c2',
-    //       ],
-    //     },
-    //     'solanadevnet',
-    //   ),
-    // },
+    solanatestnet: {
+      interval: 1,
+      reorgPeriod: getReorgPeriod(chainMetadata.solanatestnet),
+      validators: validatorsConfig(
+        {
+          [Contexts.Hyperlane]: ['0xd4ce8fa138d4e083fc0e480cca0dbfa4f5f30bd5'],
+          [Contexts.ReleaseCandidate]: [],
+          [Contexts.Neutron]: [],
+        },
+        'solanatestnet',
+      ),
+    },
+    eclipsetestnet: {
+      interval: 1,
+      reorgPeriod: getReorgPeriod(chainMetadata.eclipsetestnet),
+      validators: validatorsConfig(
+        {
+          [Contexts.Hyperlane]: ['0xf344f34abca9a444545b5295066348a0ae22dda3'],
+          [Contexts.ReleaseCandidate]: [],
+          [Contexts.Neutron]: [],
+        },
+        'eclipsetestnet',
+      ),
+    },
   };
 };