+++ typescript/infra/config/routingIsm.ts
@@ -12,18 +12,10 @@ import {
 import { DeployEnvironment } from '../src/config';
 
 import { Contexts } from './contexts';
-import { supportedChainNames as mainnet3Chains } from './environments/mainnet3/chains';
-import { owners as mainnet3Owners } from './environments/mainnet3/owners';
-import { owners as testOwners } from './environments/test/owners';
+import { environments } from './environments';
+import { ethereumChainNames as mainnet3Chains } from './environments/mainnet3/chains';
 import { supportedChainNames as testnet4Chains } from './environments/testnet4/chains';
-import { owners as testnet4Owners } from './environments/testnet4/owners';
 import { multisigIsm } from './multisigIsm';
-
-const owners = {
-  test: testOwners,
-  testnet4: testnet4Owners,
-  mainnet3: mainnet3Owners,
-};
 
 const chains = {
   test: TestChains,
@@ -55,7 +47,7 @@ export const routingIsm = (
   return {
     type: IsmType.ROUTING,
     domains: aggregationIsms,
-    owner: owners[environment][local],
+    owner: environments[environment].core[local].owner,
   };
 };