+++ typescript/infra/scripts/helloworld/kathy.ts
@@ -246,7 +246,7 @@ async function main(): Promise<boolean> {
   }
 
   chains.map((chain) =>
-    updateWalletBalanceMetricFor(app, chain, coreConfig.owners[chain]),
+    updateWalletBalanceMetricFor(app, chain, coreConfig.owners[chain].owner),
   );
 
   // Incremented each time an entire cycle has occurred
@@ -365,14 +365,16 @@ async function main(): Promise<boolean> {
       messagesSendCount.labels({ ...labels, status: 'failure' }).inc();
       errorOccurred = true;
     }
-    updateWalletBalanceMetricFor(app, origin, coreConfig.owners[origin]).catch(
-      (e) => {
-        warn('Failed to update wallet balance for chain', {
-          chain: origin,
-          err: format(e),
-        });
-      },
-    );
+    updateWalletBalanceMetricFor(
+      app,
+      origin,
+      coreConfig.owners[origin].owner,
+    ).catch((e) => {
+      warn('Failed to update wallet balance for chain', {
+        chain: origin,
+        err: format(e),
+      });
+    });
 
     // Break if we should stop sending messages
     if (await nextMessage()) {