+++ typescript/sdk/src/metadata/chainMetadataTypes.ts
@@ -111,6 +111,10 @@ export const ChainMetadataSchemaObject = z.object({
     .array(RpcUrlSchema)
     .describe('For cosmos chains only, a list of Rest API URLs')
     .optional(),
+  grpcUrls: z
+    .array(RpcUrlSchema)
+    .describe('For cosmos chains only, a list of gRPC API URLs')
+    .optional(),
   blockExplorers: z
     .array(
       z.object({
@@ -216,6 +220,20 @@ export const ChainMetadataSchema = ChainMetadataSchemaObject.refine(
       message: 'Bech32Prefix and Slip44 required for Cosmos chains',
       path: ['bech32Prefix', 'slip44'],
     },
+  )
+  .refine(
+    (metadata) => {
+      if (
+        metadata.protocol === ProtocolType.Cosmos &&
+        (!metadata.restUrls || !metadata.grpcUrls)
+      )
+        return false;
+      else return true;
+    },
+    {
+      message: 'Rest and gRPC URLs required for Cosmos chains',
+      path: ['restUrls', 'grpcUrls'],
+    },
   );
 
 export type ChainMetadata<Ext = object> = z.infer<typeof ChainMetadataSchema> &
@@ -246,4 +264,10 @@ export function getDomainId(chainMetadata: ChainMetadata): number {
 export function getChainIdNumber(chainMetadata: ChainMetadata): number {
   if (typeof chainMetadata.chainId === 'number') return chainMetadata.chainId;
   else throw new Error('ChainId is not a number, chain may be of Cosmos type');
+}
+
+export function getReorgPeriod(chainMetadata: ChainMetadata): number {
+  if (chainMetadata.blocks?.reorgPeriod !== undefined)
+    return chainMetadata.blocks.reorgPeriod;
+  else throw new Error('Chain has no reorg period');
 }