+++ typescript/infra/CHANGELOG.md
@@ -1,5 +1,34 @@
 # @hyperlane-xyz/infra
 
+## 3.6.1
+
+### Patch Changes
+
+- ae4476ad0: Bumped mantapacific reorgPeriod to 1, a reorg period in chain metadata is now required by infra.
+- e4e4f93fc: Support pausable ISM in deployer and checker
+- Updated dependencies [3c298d064]
+- Updated dependencies [ae4476ad0]
+- Updated dependencies [f3b7ddb69]
+- Updated dependencies [df24eec8b]
+- Updated dependencies [78e50e7da]
+- Updated dependencies [e4e4f93fc]
+  - @hyperlane-xyz/utils@3.6.1
+  - @hyperlane-xyz/sdk@3.6.1
+  - @hyperlane-xyz/helloworld@3.6.1
+
+## 3.6.0
+
+### Patch Changes
+
+- 67a6d971e: Added `shouldRecover` flag to deployContractFromFactory so that the `TestRecipientDeployer` can deploy new contracts if it's not the owner of the prior deployments (We were recovering the SDK artifacts which meant the deployer won't be able to set the ISM as they needed)
+- Updated dependencies [67a6d971e]
+- Updated dependencies [612d4163a]
+- Updated dependencies [0488ef31d]
+- Updated dependencies [8d8ba3f7a]
+  - @hyperlane-xyz/sdk@3.6.0
+  - @hyperlane-xyz/helloworld@3.6.0
+  - @hyperlane-xyz/utils@3.6.0
+
 ## 3.5.1
 
 ### Patch Changes